CSID-DGU#32 feat: 로그인 상태관리
