+++ frontend/package.json
@@ -17,6 +17,7 @@
     "react-icons": "^4.11.0",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
+    "recoil": "^0.7.7",
     "sass": "^1.69.1",
     "sass-loader": "^13.3.2",
     "web-vitals": "^2.1.4"
