+++ frontend/src/components/LoginModal/LoginModal.module.scss
@@ -70,17 +70,18 @@
     cursor: pointer;
     font-size: 1.5rem;
     font-size: 500;
+    font-weight: 500;
     color: var(--gray-700);
     transition: all 0.2s ease-in-out;
+
+    &:hover {
+      border: 1px solid var(--orange);
+      // font-weight: 500;
+    }
   }
 
   img {
     margin: 0 0 0 1rem;
     width: 5rem;
-  }
-
-  .button:hover {
-    border: 1px solid var(--orange);
-    font-weight: 500;
   }
 }
