+++ frontend/src/pages/MainNotice/MainNotice.module.scss
@@ -22,7 +22,7 @@
       font-size: 1.5rem;
       font-weight: 500;
       color: var(--gray-600);
-      margin-bottom: 2rem;
+      margin-bottom: 1.7rem;
     }
   }
