+++ frontend/src/pages/MyPage/MyPage.module.scss
@@ -105,7 +105,7 @@
 .eventColor {
   width: 1.2rem;
   height: 100%;
-  background: var(--gray-200, #e5e7eb);
+  background: var(--gray-200, #e5e7eb); // 이벤트 색상이 지정되지 않았을 때의 색상
 }
 
 .title {
